Only run builds for labelled pull requests
Changed the integration tests to only run on pull requests that are
labelled 'safe to test'. The trigger for pull requests is changed from
'pull_request' to 'pull_request_target' to ensure it runs in the target
environment. Combining with the check for the required label, which can
only be assigned by target repo admins will add some measure of safety.