Add secure Pr Builds test (#28)
* Only run builds for labelled pull requests

Changed the integration tests to only run on pull requests that are
labelled 'safe to test'. The trigger for pull requests is changed from
'pull_request' to 'pull_request_target' to ensure it runs in the target
environment. Combining with the check for the required label, which can
only be assigned by target repo admins will add some measure of safety.

* Add job to remove 'safe to test' label

Add a job that runs when all the other jobs have run, it will remove the
'safe to test' label.

* Add extra conditions to run the test jobs on main workflow.

* Add license info to README (#24)

* Modify the action trigger to react on pull request review submission.
Run the integrations tests when an approval.

* Change the Action trigger to pull_request_target.

Co-authored-by: Ali Asad Lotia <ali.asad.lotia@gmail.com>