Quote table name (#49)
* Quote table name

When database name contains a dash (-), BigQuery requires backticks around the name.

* Use quote_column_name to quote correctly for all databases

Also add test to ensure that quoting works

* Rename to quote_column_name to quote_identifier

The macro is also used for table identifiers, not just for column names.

* Add test for source with dashes

Quote each part separately.
This is more in line with how quoting is done by DTB, ref. https://docs.getdbt.com/reference/project-configs/quoting

* Add test for sources with identifier

The identifier property is used to denote the actual name of a source in the database, so we should use this instead of the name property when querying the database.