Fix node identifiers if test is in a package (#54)
If a (sub)package with unit tests is imported from another (parent) project, the variable project_name will be that of the parent project. We should instead use model.package_name, which is the name of the subpackage containing the unit test.

Added a subpackage with a test to ensure that this works as expected.