Include compiled query in output and trim whitespace
If we include the query in the output, it is easier for the user to understand how the test is compiled and why it fails. Since test_query isn't used, it shouldn't give any performance impact (at least not in BigQuery).

We trim whitespaces to make the query more compact and readable.